Refactor Makefile

+++ Makefile
@@ -1,20 +1,20 @@
 CWD := $(shell pwd)
 SOURCES := $(shell find source -name '*.d')
-EXAMPLES := $(shell find examples -name '*.d')
 TARGET_OS := $(shell uname -s)
 LIBS_PATH := lib
 
 .DEFAULT_GOAL := docs
 all: docs
 
+COVERAGE_TARGETS := $(shell find . -name '-tmp*.lst') $(shell find . -name '..*.lst')
+
 test:
 	dub test
-	@sh scripts/delete-junk-lst-files.sh
+	@rm -f $(COVERAGE_TARGETS)
 .PHONY: test
 
 cover: $(SOURCES)
 	dub test --coverage
-	@sh scripts/delete-junk-lst-files.sh
 
 PACKAGE_VERSION := 0.1.0
 docs/sitemap.xml: $(SOURCES)
@@ -27,13 +27,14 @@ docs: docs/sitemap.xml
 .PHONY: docs
 
 clean: clean-docs
-	rm -rf bin $(EXAMPLES)
-	rm -f -- *.lst
-	sh scripts/delete-junk-lst-files.sh
+	rm -rf bin
+	rm -f -- *.lst ..*.lst
 .PHONY: clean
 
 clean-docs:
+ifeq ($(shell test -d docs && echo -n yes),yes)
 	rm -f docs.json
 	rm -f docs/sitemap.xml docs/file_hashes.json
 	rm -rf `find docs -name '*.html'`
+endif
 .PHONY: clean-docs
